drivers/input: fp_detect: Cleanup and Simplify fp detection logic
Cleanup debug bloat and reduce extra work done by the driver for
the fingerprint detection and inline the driver as much as possible.

All of the OnePlus 7 series devices use Goodix fingerprint with
the exact same fingerprint sensor id, with hotdogg being the only
exception with a different sensor id.

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Change-Id: I3542b52883fc0330f531254f1151adfcde9d934a